Optimizing the JsonTokenizer.
This time by avoiding to copy the bytes as much as possible, and only
keeping a lineBuffer around (that we need to reference the bytes in the
token anyway), and delay all matching and checking until we know what
the token is used for.

Now the compact json (ID) format is the fastest json format of them all,
using <6 seconds to read 10000 complex messages from file.