Add testing module with advanced message matcher.
FIRST PASS ONLY.

The point of the message matcher is to have a detailed diff on mismatch.
E.g.:

expected[
    operation[4].operand to be ADD, but was DIVIDE
    operation[7].operator.number to be 5, but was 12.7
    operation[8].operator.imaginary to be Imaginary{v:4,i:1}, but was null
    operation[8].operator.number to be null, but was 14.4
]

Or for containers:

expected[
    missing value false in container.booleanSet
    unexpected value 15 in container.longSet

    missing key in container.integerMap[14] = 44
    unexpected key in container.longMap[54] = 123

    container.doubleMap[12.12] to be 987.654, but was 987.642

    // list with same value, but different order.
    container.integerSet to change orders [±0,+1,-1,±0,±0]
]

Etc.

---

Also refactore the testing stuff (again) so we have less circular
bindings.

Jackson is renamed 'core-jackson' to be clean it's a library that is
generated code against.

All core-tests and generated code tests are moved to 'testing', to
break the cyclic module dependency:
 core (test) -> generated -> core.