Handle that a union field van be of type 'void'.
Note that this is only for unions, and it should not be allowed to
declare a void field yourself. It should be used to handle the union
field value for responses if the return type is void on service methods.

```thrift
service Srv {
  void method();
}
```

Since thrift main requires that the the success union field is set (in
the wire format) even if the return type is void.

--

NOTE: This will probably make providence also require a non-null return value to
be valid, which is something I have tried to avoid. But to keep up with
compatibility, it is necessary.