Use Boolean.TRUE as void value.
both as the passed value (mostly only used for presence), both as java
instance, and using the 'true' value where some token is required. This
should give the most readable output for json and pretty serialization
for void values:

```json
[
  "test",
  "REPLY",
  44,
  {
    "success": true
  }
]
```

As the response to a void method is more readable (and understandable)
than `"success": null`.