Make CStruct and CException share builder code.
Adding common CMessageBuilder that generalizes the message builder code.
This is mainly based on the CStruct contained builder which had fixed a
number of issues that for some reason was not also fixed in the
CException builder.

Also a number of simplifications in the CUnion builder, that no longer
uses "common code" which has a number of small fixes.

And now contained messages follow the correct default value behavior.