Enum value overhaul.
Note that I decided *not* to use optional on these, as using
`Optional.ofNullable(MyEnum.findByName(theName))` is not much work. It
was mainly thought of as a generalization of the two variants, and
Optional does not yet solve the value-or-null branching problem (it may
in java 9).

- Use 'id' not 'value' for the numeric id.
- Static getters are:
    - `findById` instead of forValue
    - `findByName` instead of forName
    - `valueForId` non-null vasriant of findById
    - `valueForName` non-null variant of findByName
- `getId` and `getName` are default impl in PEnumBalue, and just wraps
  `asInteger`, `asString`.
- `getValue` is deprecated and wraps `asInteger`.

Enum descriptor changes

- getValues() is @Nonnull
- Getters are changed to match the static getters. (legacy methods
  removed).

Also:

- Added lots of @Nullable and @Nonnull annotations.
- Moved pregen classes in providence-reflect to proper namespace.

PS: The immense number of changes lines comes from the fact that
commonly used methods have changed name.