Create a rolling and queue file message writer.
the rolling file message writer writes to files the same fashion as a
logging system's rolling file writers.

The queue writer is meant to be used to funnel many threads write
operations into a single thread that writes to the actual file. This is
important to avoid IO contentions, and to make the write operations
as fast as possible to be able to not block other important threads,
that just want something logged.