Update morimekta IO utils, and fix json.
The JSON tokenizer biggest functional change is that if there is an
ending newline after the JSON struct, it will be consumed by the
tokenizer. This will force a parallell update of utils and providence
where JSON streams are used.

Updating IT testing for serialization with running parallell threads,
even though each serialization and deserialization is not handled
through queueing etc, it is closer to the reality of how the libraries
will be used (e.g. in jetty, file IO etc).

NOTE: In TJSONProtocol, there appears to be a newline within the stream
which mnesses up deserialization. This made me have to remove the
TJsonProtocol from the IT testing. The bug seems to have appeared when I
updated to newer jfairy (more varied unicode chars), which may be the
cause of the bug. TODO: Figure exactly what goes on and fix?